ci: Build a windows docker image and publish it to the registry
This change allows to build a windows docker image of `sfcgal` and
publish it to the registry.
The docker image is built on 3 different occasions:
- a new Merge Request (manual, does not block the next stages)
- a new tag (automatic)
- a push to master (automatic)

The merge request job one is manual because this CI is quite
slow. Indeed, windows runners are still in beta and the VMs take some
time to start (up to 5 minutes) and they do not have a lot of compute
ressources. According to gitlab tickets, the situation should improve
in Q4 2024.

The `dockerfile` uses a base docker image which already contains
`sfcgal` dependencies built with vcpkg (mostly boost and
cgal). Indeed, windows base image provided by gitlab already contain
`vcpkg`, `cmake` and visual studio c++ compiler but compiling `boost`
and `cgal` with `vcpkg` can take longer than the runner timeout (2h).

These images will be used by `pysfcgal` to build a windows wheel.