fix(wkb): add support for Solid and MultiSolid with correct WKB codes 25/26
Previously, the WKB export did not support `Solid` and `MultiSolid`, as these
geometries are not commonly found in existing standards. However, WKT export
was already implemented for both types. According to the ISO/IEC 13249-3:2016
standard, the code 102 is assigned to `AffinePlacement`, and while `ST_SOLID`
is not instantiable, `SOLID` is the name used in (City)GML, corresponding to
`ST_BRepSolid`.

Our implementation of `Solid` aligns with `ST_BRepSolid`, which consists of
shells (represented as `PolyhedralSurface` in our case;
`PolyhedralSurface` + `CompoundSurface` in the standard).
However, the WKB code is not officially listed for `BRepSolid`, except for
`BRepSolidZ` with the code `<wkbbrepsolidz> 1025`.

This implies that the regular code for `BRepSolid` should be 25
(with 1000 added for the Z coordinate). Additionally, since `MultiSolid` has no
existing WKB code, we propose assigning the code 26, as it is currently
available.
Thus, we move from the previous 101/102 to 25/26, ensuring we use the correct
WKB code for `Solid` (BRepSolid).