ci: Pages and build:documentation need to have the same limitations
The 'pages' depends on the 'build:documentation' one. However, the
build one has some limitations that the 'pages' one does not have
one. Therefore, the 'deploy' job is always called on changes to master
while 'build:documentation' has not been called because of its
"changes" limitations. This will result in a ci error.

This issue is fixed by ensuring that the 'build:documentation' and
'pages' jobs have the same limitations.