docker: Ignore the vcpkg.json to hide it from gitlab CI
`vcpkg.json` is supposed to be used by the github CI to call `vcpkg
install` to install SFCGAL dependencies. However, it is
also also by the gitlab CI. But the gitlab base image already has the
SFCGAL dependencies. By adding it to the `.dockerignore`, it prevents
using it in the Gitlab CI.