fix(capi): Fix deprecated usage
The `deprecated` or `__deprecated__` keywords are not handled by all
the versions of the compilers. For example, the current version of
clang in debian stable does not handle it.

In order to fix this issue, provide a fallback mechanism by
introducing a "SFCGAL_DEPRECATED" macro which should work in all
cases.

See for example: https://gitlab.com/sfcgal/SFCGAL/-/jobs/9501302274
Fixes: de89601144a5c82c02314757e77663236e3a4f04