chore: add commitizen to the project
This new tool is added to the pre-commit config, the CI (for each
merge request) and mentionned in the documentation.

Plus, the release process is now briefly documented. To be noted:
commitizen allows to automatically generate a CHANGELOG diff,
specifically before any version upgrade:

```bash
cz changelog --incremental --unreleased-version <new-tag>
```

Disclaimer: the commit must be written following the commitizen
guidelines so as to be considered by commitizen. A manual check will
have to be done at the first release after this commit.