chore: Update clang-format to version 19
Version 16 is really old at this point and there has been a lot of
changes between version 16 and 17. This ensures to keep an up to date
version.

`kszonek/clang-format` docker image latest version is 16. To solve it,
switch to a debian stable slim image and install `clang-tidy-19` on
it.

On the `pre-commit` side, `ssciwr/clang-format-hook` is also
outdated. Switch to `pre-commit/mirrors-clang-format` which provides
newer versions.