fix(getpointsvisitor): Use patch getters for TriangulatedSurface
a TriangulatedSurface is a not a GeometryCollection and it should be
treated as one geometry.

- `TriangulatedSurface::numGeometries()` should return 1 and
`TriangulatedSurface::numPatchs()` the number of triangles
- `TriangulatedSurface::geometryN()` should return this and
`TriangulatedSurface::patchN()` the internal triangles

This behavior will be changed in the following commits. This commit
prepares this change on the `WkbWriter` side by calling
numPatchs/triangleN for TriangulatedSurface instead of
numGeometries/geometryN.

See:
https://lists.osgeo.org/pipermail/postgis-devel/2025-April/030517.html