fix(distance3d): Do not treat PolyhedralSurface as a collection
a PolyhedralSurface is a not a GeometryCollection and it should be
treated as one geometry.

- `PolyhedralSurface::numGeometries()` should return 1 and
`PolyhedralSurface::numPatchs()` the number of polygons
- `PolyhedralSurface::geometryN()` should return this and
`PolyhedralSurface::patchs()` the internal polygons

This behavior will be changed in the following commits. This commit
prepares this change on the `distance3D` side by calling
numPatchs/patchN for PolyhedralSurface instead of
numGeometries/geometryN.

See:
https://lists.osgeo.org/pipermail/postgis-devel/2025-April/030517.html