chore(build): Fix compilation with boost 1.89
In the upcoming Boost 1.89.0 release, Boost.System stub has been
removed which causes CMake error:

```
CMake Error at /opt/homebrew/lib/cmake/Boost-1.89.0/BoostConfig.cmake:141 (find_package):
  Could not find a package configuration file provided by "boost_system"
  (requested version 1.89.0) with any of the following names:

    boost_systemConfig.cmake
    boost_system-config.cmake
```

This issue is fixed by dropping `system` from the `COMPONENTS` as it's
not needed since 1.69. With this change, boost minimum version to
build SFCGAL is now 1.69.0. This should not be an issue as this
version was released in 2018.

Closes: https://gitlab.com/sfcgal/SFCGAL/-/issues/306