fix(primitive): Add a logic to validate parameters compatible with c api
Each primitive has its own setter and checks the validity of its
parameter. For example `Torus::setMainRadius` checks that the new main
radius is positive and throws and exception if that's not the case.

On the other hand, there is also `Primitive::setParameter(name,
value)` which checks that:
- parameter `name` exists in the primitive
- the type of the value is compatible with the parameter

However, it does not check the validity of the parameter. If a radius
is positive, for example. `setParameter` is mostly used by the C
API. This means, that there is no way to check the validity of the
parameter when using the C API.
This issue is fixed by introducing 2 new methods:
- `validateParameters`
- `validateAndSetParameter`

Both are protected. The first one is supposed to be implemented by
each child class. One finds inside all the constraints that the
parameters must validate (for example a radius must be positive). It
works on a copy of `m_parameters` to prevent setting wrong values
before validating them. The second one calls `validateAndSetParameter`
to check the parameters and sets the new value if the constraints are
respected. Then, it calls `invalidateCache` to clear the cache.

This implies two changes:
- on the c++, each setter now simply directly calls
  `validateAndSetParameter`
- on the C API side, the previous issue is fixed because
  `setParameter` now calls `validateAndSetParameter` to add the
  validation step