ci(doxygen): Ensure to always add the corresponding header files
For most of the '.cpp' or '.c' files, the documentation is provided in
the '.h' file. If only the 'cpp' or 'c' file is modified, the doxygen
check will return errors because it won't check the corresponding
header files.

This issue is fixed by adding the corresponding header files if it
exists and it's not already part of the changed_files list.