feat: Introduce DereferenceIterator
This will be used in the next commits to simplify of
`std::vector<std::unique_ptr<Geometry>>`. Indeed, when directly using
an iterator such as `container.begin()`, one get a `std::unique_ptr`,
which is cumbersome to use with the iterator logic.

This can be simplified by wrapping the iterators so that the elements
are dereferenced in the iterator. This way, one get direct access to
the elements instead of `std::unique_ptr`. This is exactly what
`DereferenceIterator` does.

In pratice, one simply need to create an iterator with
`dereference_iterator`:

```
dereference_iterator(container_of_unique_ptr.begin());
```

See: https://jonasdevlieghere.com/post/containers-of-unique-pointers/